Add: Function handler to get code from my plugin
So the basic plan for now is to send the code in my current buffer to the backend server (this). Once i have the code, I can use openAI api or codellama to get code completions. This could serve as a middleware microservice. This seems to be a useless way of getting response from llm but for now this is the fix. XD